+++ libraries/core-react/src/components/TextField/TextField.tsx
@@ -9,7 +9,7 @@ import styled, { ThemeProvider } from 'styled-components'
 import { Field } from './Field'
 import { Label } from '../Label'
 import { HelperText } from './HelperText'
-import { TextFieldProvider } from './context'
+import { TextFieldProvider } from './TextField.context'
 import type { Variants } from './types'
 import { textfield as tokens } from './TextField.tokens'
 import { useToken } from '../../hooks'
